# spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "tinistreamer",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Health check",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/info": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Get info",
        "description": "Get information about the server",
        "operationId": "get_info",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/client/sse": {
      "get": {
        "tags": [
          "Client"
        ],
        "summary": "Connect SSE stream",
        "description": "Connect to a stream and receive SSE events",
        "operationId": "connect_sse",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "uint8",
                    "minimum": 0.0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ClientToken": []
          }
        ]
      }
    },
    "/api/stream/": {
      "get": {
        "tags": [
          "Stream"
        ],
        "summary": "List streams",
        "description": "List all active streams",
        "operationId": "list_streams",
        "parameters": [
          {
            "name": "pattern",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StreamInfo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      },
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Create stream",
        "description": "Create a new stream, and get a client URL and token to connect to the stream",
        "operationId": "create_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamAccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/stream/token": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Create stream token",
        "description": "Create a new client token to connect to a stream",
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamAccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/stream/add": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Add events",
        "description": "Add events to a stream",
        "operationId": "add_events",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEventsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddEventsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/stream/add/json-stream": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Add events JSON stream",
        "description": "Add events to a stream via a JSON stream",
        "operationId": "add_events_json_stream",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "JSON stream (stream of JSON strings separated by newlines)",
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddEventsStreamResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/stream/cancel": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Cancel stream",
        "operationId": "cancel_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndStreamResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/stream/end": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "End stream",
        "operationId": "end_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndStreamResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Incorrectly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "InfoResponse": {
        "type": "object",
        "required": [
          "redis",
          "url",
          "version"
        ],
        "properties": {
          "url": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "redis": {
            "$ref": "#/components/schemas/RedisStats"
          }
        }
      },
      "RedisStats": {
        "type": "object",
        "required": [
          "static",
          "streaming",
          "streaming_available",
          "streaming_in_use",
          "streaming_max"
        ],
        "properties": {
          "static": {
            "description": "Number of static connections",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "streaming": {
            "description": "Number of streaming connections",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "streaming_in_use": {
            "description": "Number of in-use streaming connections",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "streaming_available": {
            "description": "Number of available streaming connections",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "streaming_max": {
            "description": "Maximum number of streaming connections",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "StreamInfo": {
        "description": "Information about the stream",
        "type": "object",
        "required": [
          "key",
          "length",
          "ttl"
        ],
        "properties": {
          "key": {
            "description": "Key of the stream in Redis",
            "type": "string"
          },
          "length": {
            "description": "Number of events in the stream",
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0
          },
          "ttl": {
            "description": "Expiration of the stream",
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "StreamAccessResponse": {
        "type": "object",
        "required": [
          "token",
          "url"
        ],
        "properties": {
          "url": {
            "description": "URL to connect to the stream",
            "type": "string"
          },
          "token": {
            "description": "Bearer token to access the stream",
            "type": "string"
          }
        }
      },
      "StreamRequest": {
        "type": "object",
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string"
          }
        }
      },
      "AddEventsResponse": {
        "type": "object",
        "required": [
          "ids"
        ],
        "properties": {
          "ids": {
            "description": "IDs of the added events",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AddEventsRequest": {
        "type": "object",
        "required": [
          "events",
          "key"
        ],
        "properties": {
          "key": {
            "description": "Key of the stream to write to",
            "type": "string"
          },
          "events": {
            "description": "Events to add to the stream",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddEvent"
            }
          }
        }
      },
      "AddEvent": {
        "type": "object",
        "required": [
          "event"
        ],
        "properties": {
          "event": {
            "description": "Name/type of the event",
            "type": "string"
          },
          "data": {
            "description": "Event data",
            "type": "string",
            "nullable": true
          }
        }
      },
      "AddEventsStreamResponse": {
        "type": "object",
        "required": [
          "errors",
          "ids"
        ],
        "properties": {
          "ids": {
            "description": "IDs of the added events",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "errors": {
            "description": "Errors that occurred while adding events",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EndStreamResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "description": "ID of the ending event",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Provide backend API key.",
        "type": "apiKey",
        "name": "X-API-KEY",
        "in": "header"
      },
      "ClientToken": {
        "description": "Provide client token as a Bearer token.",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "bearer"
      }
    }
  }
}